Addresses issue #1349
There is an edge use case where samples aren't collected from workers
because `num_samples` have already been collected from some subset of
the sampler workers, then the other workers should have their calls
to rollout canceled. At the moment, the fix for this is to only call
rollout on workers if rollout should be called on them. For example
If there are 2 workers, and they are supposed to each collect 16
samples and then a call to obtain samples is made asking for only 16
samples, then only the first sampler worker will be started. The best
way for us to handle in the future is to shutdown workers using some
ipc, but handling starting workers the way that is mentioned above
is the best fix for a backport.